refactor(docs): modal dx
